Update torch version 1.3.1 to 1.6.0 (#274)
* Update torch version 1.3.1 to 1.6.0

In R2D1 brain.py specify tensor type to use .float().
In PPO, seperate loss that reapeated in training phase.
In SAC, Fix to use updated Q-networks in training actor phase.

* Delete with torch.autograd.set_detect_anomaly(True)

* Change position of training Q,V fucnction temrs to below.

* Change position of training Q,V function to below

* Fix requirements.txt to update torch 1.6.0 (up-to-date)

* Specify torch version 1.6.0 in requirements.txt

* Fix pytorch 1.6.0 & pylint error(w0223)

* Upgrade pylint and pytest-flake8 to solve conflicting problem

+++ rl_algorithms/common/networks/brain.py
@@ -126,7 +126,7 @@ class GRUBrain(Brain):
         hidden = torch.transpose(hidden, 0, 1)
 
         # Unroll gru
-        gru_out, hidden = self.gru(gru_input, hidden)
+        gru_out, hidden = self.gru(gru_input.float(), hidden.float())
 
         # Get q
         q = self.head(gru_out.contiguous().view(batch_len * seq_len, -1))
